Source code for previous test + new exception file + further tests
Coded the functionality of the program to throw an exception whenever the rgb value input in the palette is invalid. Exception Handling file is separate

Also coded test that checks if the program properly throws an exception when a DUPLICATE color is input into our palette, as we were told we need to handle this evet somehow.