src code for previous test / new exception file / new test
Created code that handles the possibility of adding a 'duplicate' color in the palette by throwing an exception. The exception handling code is written in a separate file.

Added testing code that handles the exception for when an attempt to add a new color is made to the palette, but the size is already full.